backend/pyproject.toml
[project]
name = "vc-trading-cards-backend"
version = "0.1.0"
description = "FastAPI backend: Discord OIDC + ACA-Py multitenancy for Tritone Cards (The Devil's Interval Collectible Cards)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.29.0",
    "discord-interactions>=0.4.0",
    "webauthn>=2.7.1",
    "Pillow>=10.0.0",
    "easyocr>=1.7.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
